made default dirs visible from outside
Suppose a user wants to redefine how data files are searched, but would
like to fall back to the paths in CamomileDefaultConfig.  If the user
tries to access those paths thru CamomileLibraryDefault.Config, then
the program will crash if the paths in CamomileDefaultConfig are wrong,
*even if the custom search procedure would otherwise succeed*. This is
because the module CamomileLibraryDefault also evaluates a functor, and
that fails. (Same thing with CamomileLibraryDyn.)