Camomile/locales/eo.txt: Fix license by importing data from ICU
The license for this particular (Esperanto) file is non-free.
Possibly it is supposed to be overridden by the "license.html" file in
the current directory, but our tooling can't read HTML files.

We can fix this by importing the same data from the newer Unicode ICU
project files, as these are distributed under a free license.  This
also gives us slightly increased functionality.

See also:

https://lists.fedoraproject.org/archives/list/legal@lists.fedoraproject.org/thread/EP453A4JSXM2DKIBLPYBKLID47OCKCPW/

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>